xulrunner: boneyard
Mozilla have dropped support for xulrunner and currently no developer has picked it up.

On our side of the deal, beyond this summer when a new major release of OS X is
released we'll lose the ability to build & test this formula completely as it
currently doesn't build on anything newer than Mavericks.

Ref: https://lists.mozilla.org/pipermail/dev-platform/2015-July/011178.html
Ref: https://bugzilla.mozilla.org/show_bug.cgi?id=1221724